clean up and test cases
