+++ .gitignore
@@ -2,11 +2,6 @@ env
 *.pyc
 *.DS_Store
 staticfiles
-dataload.sh
-us-ca-solano_county*
-TODO.txt
-data/*
 static/libs
-dev_prod.py
-intake/import_rms.py
-intake/import_crw.py
+local_prod.py
+temp_utils
