--- data_load/templates/data_load/import_csv.html
@@ -1,6 +1,0 @@
-<form enctype="multipart/form-data" method="post" action="/data_load/import_csv/">
-{%csrf_token %}
-<table>
-{{ form.as_table }}
-</table>
-<input type="submit" value="Submit" />
