+++ intake/utils.py
@@ -1,10 +1,11 @@
 # TODO: abstract out logic from views to here
 
+
 def create_call(call_sid):
 
     pass
 
+
 def update_call(call_sid, post_params, param_key, model_field):
 
     pass
-
