--- workflow/templates/workflow/call_audit_log.html
@@ -1,41 +1,0 @@
-{% extends "base/layout.html" %}
-
-{% load staticfiles %}
-
-{% block assets %}
-
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
-
-<link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css" rel="stylesheet">
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js"></script>
-
-<link href="https://cdn.datatables.net/1.10.7/css/jquery.dataTables.css" rel="stylesheet">
-<script type="text/javascript" charset="utf8" src="//cdn.datatables.net/1.10.7/js/jquery.dataTables.js"></script>
-
-<link rel="stylesheet" href="{% static "base/base.css" %}"/>
-<link rel="stylesheet" href="{% static "workflow/css/table_page.css" %}"/>
-<script src="{% static "workflow/js/call_audit_log.js" %}"></script>
-{% endblock %}
-
-{% block body %}
-<div class="container">
-
-  <div class="row">
-    <div class="col-md-12">
-        <table id="data-table" class="table table-hover row-border">
-          <thead>
-            <tr>
-                <th>Call Time</th>
-                <th>Change Timestamp</th>
-                <th>User</th>
-                <th>Changed Field</th>
-                <th>Original Value</th>
-                <th>Updated Value</th>
-            </tr>
-          </thead>
-        </table>
-    </div>
-  </div>
-  
-</div>
-{% endblock %}
