+++ workflow/templates/workflow/landing.html
@@ -82,7 +82,6 @@
             <div id="map">
                 <script src="{% static "workflow/js/landing.js" %}"></script>
             </div>
-            <a href="/workflow/map">view the full map here</a>
         </div>
 
         <div class="col-md-4">
