+++ workflow/templates/workflow/map.html
@@ -11,9 +11,6 @@
 <script src='https://api.mapbox.com/mapbox.js/v2.2.1/mapbox.js'></script>
 <link href='https://api.mapbox.com/mapbox.js/v2.2.1/mapbox.css' rel='stylesheet' />
 
-<link rel="stylesheet" href="{% static "base/bootstrap-switch.min.css" %}"/>
-<script src="{% static "base/bootstrap-switch.min.js" %}"></script>
-
 <link rel="stylesheet" href="{% static "base/base.css" %}"/>
 <link rel="stylesheet" href="{% static "workflow/css/map.css" %}"/>
 <script src="{% static "workflow/js/map.js" %}"></script>
