+++ workflow/urls.py
@@ -1,0 +1,20 @@
+from django.conf.urls import patterns, url
+
+urlpatterns = patterns(
+    '',
+    # case related views
+    url(r'^workflow/case/(?P<case_id>\d*)/$', 'workflow.views.case_views.case', name='case'),
+    url(r'^workflow/cases/$', 'workflow.views.case_views.cases', name='cases'),
+    url(r'^workflow/cases_data/$', 'workflow.views.case_views.cases_data', name='cases_data'),
+    url(r'^workflow/add_case_assignee/$', 'workflow.views.case_views.add_case_assignee', name='add_case_assignee'),
+    url(r'^workflow/remove_case_assignee/$', 'workflow.views.case_views.remove_case_assignee', name='remove_case_assignee'),
+
+    # verification related views
+    url(r'^workflow/verification/(?P<verification_id>\d*)/$', 'workflow.views.verification_views.verification', name='verfication'),
+
+    # report related views
+    url(r'^workflow/report/(?P<report_id>\d*)/$', 'workflow.views.report_views.report', name='report'),
+    url(r'^workflow/reports/$', 'workflow.views.report_views.reports', name='reports'),
+    url(r'^workflow/reports_data/$', 'workflow.views.report_views.reports_data', name='reports_data'),
+    url(r'^workflow/add_report/$', 'workflow.views.report_views.add_report', name='add_report'),
+)
